Improve documentation of SoftmaxOutput
This change improves the documentation of the `SoftmaxOutput` symbol by
adding information regarding the calculated cross entropy loss. The old
documentation only contained information on the softmax function. The
documentation was also wrong in the sense that the softmax function has
no label information as part of its input. It only turns a vector into a
probability vector. The changed documentation explains the softmax
function and the cross entropy loss on their own and their combination
to describe the calculation represented by the `SoftmaxOutput` symbol.