Add script to download MNIST handwritten digit dataset
The MNIST handwritten digit database can be used as test data for neural
networks and other machine learning algorithms.  We will be using it in
particular to provide integration tests for the MXNet D wrapper library.

This script will search for files providing the MNIST handwritten digit
database, downloading them from the project website if not found.  All
files will be compared to known SHA256 checksums to ensure they are as
expected, whether they have been newly downloaded or already exist on
the local system.

By default, data will be downloaded to `${HOME}/.cache/mnist`, but the
user can specify an alternative location using the `MNIST_DATA_DIR`
environment variable or simply as an argument to the script, e.g.:

    script/download-mnist /var/cache/mnist

or

    MNIST_DATA_DIR=/var/cache/mnist script/download-mnist