Add integration test suite using MNIST dataset
This simple test suite performs multi-class logistic regression on the
MNIST dataset using the D MXNet wrapper library.  It does not provide
comprehensive coverage (and is not intended to), but should provide a
basic sanity check of the library's behaviour.

A build dependency on `download-mnist` is included to ensure the dataset
will be available before the test suite is run.  An extra Travis install
stage has been added to ensure that CI will download the data in advance
of invoking `make`, meaning that download failure will count as an error
rather than a build failure.  Travis cache settings have also been added
to prevent the data needing to be downloaded again and again with every
individual build.  Since `beaver` automatically ensures that `$HOME` is
shared between the host system and the docker container, we do not need
to add any custom setup to provide Travis' caching service with access
to the data download directory.

`LDFLAGS` have been extended for the `test-mxnet.stamp` target so as to
include libraries required by the `ocean` functionality the test suite
depends on.  The docker build script has been updated to install these
dependencies.