Add debugging help when freeing handles in the destructor
Freeing non-null handles in the destructor is considered to be a
programming error because they should be freed by an explicit call to
`freeHandle`. This change adds debugging output when a non-null handle
is freed inside the destructor. Because relying on the GC for freeing a
handle in time is not reliable. It is reliable for `delete` or `destroy`
but we prefer an explicit call to `freeHandle` prior to a manual
`delete` or `destroy`. The policy is to always free the handle manually
using `freeHandle`.

This debugging diagnostics is enabled for `MXNetHandleManualFree` debug
builds.