Add function for writing a formatted string for debugging
This change introduces a function to write to standard error which won't
perform any allocation unless a string conversion of an argument
requires such. As such it can be used for writing errors in contexts
where no allocations are allowed. Most notably this includes object
destruction. This function will be used in upcoming changes to replace
error reporting that may be executed during destruction. The function is
supposed to be used for error reporting inside `debug`.