Refactor error reporting used for debugging
In five places errors are reported which may run in the context of
object destruction. Since it is not allowed to allocate during
destruction a buffer together with `snformat` is used in four of those
places. The other case uses `sformat` and a sink which is identical to
the implementation of `debugReport`. This change replaces this error
reporting by calling the recently introduced function `debugReport`
which also performs no allocation for these cases. This patch also
changes two usages of `Stderr.formatln` inside `debug` for consistency.