Add Build/Use, Releases, and Contributing sections to README file
The `Build/Use` section describes the dependencies of the library, with
package and submodule dependencies covered in separate subsections.
Notes on how to build, run tests, and convert to D2 are also provided
(the latter replacing the original section on D2 compatibility).

The `Releases` and `Contributing` sections are standardized sections for
Neptune libraries: the first describes where to find library releases
and the structure of the release notes, while the second provides a link
to contributor documentation (which includes release-notes management).