Call `outputs` in executor unittests only once
Since `outputs` returns n-dimensional arrays referencing the same data
for multiple invocation of `outputs` it is sufficient to call it only
once. It also avoids freeing any existing `NDArray` handle set by a
previous call just to set it to a new handle referencing the same
underlying elements. Only `NDArray` instances with `null` handles are
passed to `outputs`.
This change updates the unittests to call outputs only once. This
does not change the semantics of tests but makes only as many calls as
necessary.