Clarify documentation for executor `outputs` method
This change extends the documentation of the `outputs` method for the
`Executor`. It adds that the outputted n-dimensional arrays share the
underlying data over multiple calls. Hence, there is usually no need to
call the `outputs` method more than once. It also adds that any
underlying handle of passed n-dimensional arrays is replaced, i.e.,
with any pre-existing handles being freed.
Also the `Returns` section is removed as this method does not have a
return parameter and in the `Throws` section the type of the thrown
exception is corrected to `MXNetExecutorException` and a typo regarding
the ordering of words is fixed.
The `unittest` of the method is updated to verify the extended
documentation.