Require handles to be null when providing the executor outputs
This change introduces the requirement to pass in output `NDArray`
instances with `null` handles. Previously, non-`null` handles were
allowed which can be surprising to the caller when freed under the hood.
Note, this is a breaking change. But it is better to force the caller to
be explicit about the `NDArray` instanced updated with an executor's
outputs. The documentation of the method is updated accordingly and the
`Throws` section is updated to include the case of non-`null` handles.