Replace all usages of assigning to `data` with `copyFrom`
This change uses `copyFrom` to assign data to n-dimensional arrays in
the tests. With `copyFrom` it is guaranteed that the old data has been
read by pending reads before performing the writes. The write access is
synchronized with pending reads. There may be cases where you don't need
this synchronization (because you know there are no pending reads) but
in general and especially for testing it is better to be on the safe
side here. `copyFrom` avoids data race issues for multi-threaded MXNet
engines.

Part of sociomantic-tsunami/dmxnet#53.