Set MXNET_ENGINE_TYPE=NaiveEngine in Config.mak if not already set
This should provide a sensible default engine for local testing, which
which may readily be overridden by the user, e.g.:

    MXNET_ENGINE_TYPE=ThreadedEngine make test all

It should not interfere with cases where `MXNET_ENGINE_TYPE` is already
set in the build environment.

As a friendly by-product, since this tweak ensures `MXNET_ENGINE_TYPE`
is always set as an environment variable, libmxnet should now always log
the choice when running tests.