Add closure allocation check to CI stages
This should allow us to verify that we are not accidentally generating
garbage via the use of closures.  The new `closures.sh` ci helper shell
script exports `DFLAGS=-vgc` to make the compiler list GC allocations,
then invokes the `fasttest` target and greps its output for `closure`.

`set -o pipefail` is used to ensure that non-zero `make` exit statuses
propagate, and the exit status of the `grep` call is checked to ensure
that it is 1 (i.e. that `grep` did not find any `closure`, but exited
without error).

Note that in principle `make` can generate an exit status of 1, but only
if the `-q` option is used, which should never be the case, so we should
be safe relying on it exiting either with 0 (success) or 2 (error).  We
should not need to care about it in any case, since `grep`'s exit status
should be checked separately thanks to the enclosing parentheses.

This should ensure that the test case will only pass if both the D build
succeeds and there are no reports of closure allocation.

Note also that we cannot use `set -e` here as this would fail for `grep`
exit status 1: we have to use manual exit status checks.