Ensure MNIST usage example does not require loading the dataset
The integration-test MNIST usage example should not require us to really
load the dataset, as this makes an implicit assumption that the dataset
has already been downloaded.  By using a nested function declaration, we
can check the validity of the example code without actually needing to
run it.

This should prevent unittest failures on systems where the MNIST dataset
has not already been downloaded.