Remove inappropriate d1to2fix usage of (&this)
Given a `struct T`, `typeof(this)` will be `T*` in D1 but `T` in D2.
For this reason d1to2fix converts `this` to `(&this)` when used inside a
struct context, as this allows code to keep working with D1 usages such
as `*this`.

However, now that dmxnet has converted to D2 only, all these uses can be
removed: for example, `*(&this)` (converted from D1 `*this`) is just a
noisy way of writing `this`, while the remaining uses are all cases of
`(&this).some_field` in which the practical impact is a no-op (since the
pointer is dereferenced anyway when looking up the field).

This patch replaces all such unnecessary `(&this)` occurrences in the
code.