Drop unnecessary parentheses for const from the C bindings
For cases like `const(int)` the parentheses are not necessary. This
change removes the parentheses in these cases from the C bindings.
Note for cases like `const(char)*` parentheses are necessary because
`const` is greedy, i.e., `const char*` is equivalent to `const(char*)`.

Also dropping the parentheses in cases where they are not needed brings
us closer to the output of dstep which simplifies generating the
bindings in future.