Switch to equal based alias syntax for bindings
The bindings used to use the syntax `alias name new_name` to
introduce new aliases. I believe this was because the D1 compiler only
supported this syntax. With the migration to D2 we can use the syntax
`alias new_name = name`. This change adapts the bindings to use the new
syntax exclusively. The nice side-effect is that this syntax is
generated by dstep by default which frees us from performing these
manual changes any longer.